Week-9x-Docker-Swarm/backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
ultralytics>=8.0.0
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
pydantic>=2.0.0
